BUG FIXES
output files are now open unbuffered
fixed error when evaluating stats for very large file objects
increased MaxQueue to significantly reduce the statistical probability
of all threads processing the contents of a directory and at the same
time the input queue filling so everything becomes log jammed. This
will be revised for a comprehensive fix.